Add Rust OpenCL version.

+++ week_03/heat_stencil/heat_stencil/src/print_temperature.rs
@@ -1,4 +1,4 @@
-pub fn print_temperature(matrix: &Vec<f64>, n: usize, m: usize) {
+pub fn print_temperature(matrix: &Vec<f32>, n: usize, m: usize) {
   let colors = [' ', '.', '-', ':', '=', '+', '*', '#', '%', '@'];
 
   let max = 273.0 + 30.0;
@@ -33,8 +33,8 @@ pub fn print_temperature(matrix: &Vec<f64>, n: usize, m: usize) {
 
       let temp = max_t;
 
-      let mut c = ((temp - min) / (max - min)) * colors.len() as f64;
-      c = if c >= colors.len() as f64 { colors.len() as f64 - 1.0 } else if c < 0.0 { 0.0 } else { c };
+      let mut c = ((temp - min) / (max - min)) * colors.len() as f32;
+      c = if c >= colors.len() as f32 { colors.len() as f32 - 1.0 } else if c < 0.0 { 0.0 } else { c };
 
       print!("{}", colors[c as usize]);
     }
